Enable comma-like search for nix apps
To search among nix desktop applications currently not installed, start
by typing a comma `,` character, and launch will use `nix-index` to find
all available desktop apps.

This currently guesses the name of the executable. This could be
improved by parsing the .desktop file in the nix-shell (we cannot parse
it the moment `nix-locate` returns it to us, because at that point it
might not be installed yet.